Add bottleneck & flox pkgs to user env
bottleneck probably should have been in envs all along; it's a "standard" xarray
speed-up.

xarray-2022.06.0 brought indexing refactoring and use of flox for groupby
operations. Since resample is a special case of groupby, it is to be hoped that
flox will have a benefit there too.

bottleneck and flox have been successfully tested in dev, CI test, and Doug's
user envs for the past month.