Add `output model coords` option to extraction
Set an extraction configuration item `output model coords: True` to cause the
extracted dataset to use model coordinate names instead of the defaults
(time, depth, gridY, gridX). This is primarily intended for use by the
SalishSeaNowcast make_averaged_dataset worker which needs to be able to produce
day-averaged datasets that have the same coordinate names as the hour-averaged
ones that they are down-sampled from.