Remove user-supplied cluster handling in `reshapr info` (#98)
Remove the code that allowed a user to supply a cluster description to
`reshapr info`. With no way to distinguish between a cluster description YAML
file and a model profile YAML file, we opt to handle the user-supplied model
profile because that is the more important and common use case.

In addition, handle a previously unhandled situation where we attempt to read
the description from a model profile but the 'description' key is not present.
Now, if this key is not found, an error message will be logged. That is also the
result if a user-supplied cluster description file is given to `reshapr info`.