Change to use "MS" for monthly resampling frequency
Change the resampling frequency for monthly resampling to "MS" (string alias for
`pandas.tseries.offsets.MonthBegin`) if it ends with "M" (string alias for
`pandas.tseries.offsets.MonthEnd`) to ensure that time coordinate values are set
as the 1st days of the months, which later can be shifted to the middle of the
months. This relieves the users from having to delve into the details of pandas
date offset frequency string aliases.