Correctly calculate time coordinate values in resampled extractions (#105)
* Change to use "MS" for monthly resampling frequency

Change the resampling frequency for monthly resampling to "MS" (string alias for
`pandas.tseries.offsets.MonthBegin`) if it ends with "M" (string alias for
`pandas.tseries.offsets.MonthEnd`) to ensure that time coordinate values are set
as the 1st days of the months, which later can be shifted to the middle of the
months. This relieves the users from having to delve into the details of pandas
date offset frequency string aliases.

* Refactor calc of resampled dataset time coord values

Add the `_calc_resampled_time_coord()` function that uses
`pandas.tseries.frequencies.to_offset()` to correctly calculate offsets to move
time coordinate values to the middle of resampling time intervals for both
monthly and non-monthly resampling, and return a new `pandas.DatetimeIndex` for
the resampled dataset. This streamlines the calculation of resampled dataset
time coordinates by using a more succinct pandas-based solution. The unit tests
cover various time frequency resampling scenarios to ensure that the output is
as expected. Overall, this is a more maintainable and testable way to handle
resampled dataset time coordinate value calculations.