Add `yyyymm01()` date formatter function
Added `yyyymm01()` date formatter which returns the first day of the month. It
is integrated into the `nc_files_pattern.format()` function in the data paths
list calculator. The changes are part of the support for extractions from
month-averaged datasets. Unit tests were also implemented to ensure the correct
functionality of the new formatter.