Add extraction from month-averaged datasets (#109)
* Add `yyyymm01()` date formatter function

Added `yyyymm01()` date formatter which returns the first day of the month. It
is integrated into the `nc_files_pattern.format()` function in the data paths
list calculator. The changes are part of the support for extractions from
month-averaged datasets. Unit tests were also implemented to ensure the correct
functionality of the new formatter.

* Add `yyyymm_end()` date formatter function

Added `yyyymm_end()` date formatter which returns the last day of the month. It
is integrated into the `nc_files_pattern.format()` function in the data paths
list calculator. The changes are part of the support for extractions from
month-averaged datasets. Unit tests were also implemented to ensure the correct
functionality of the new formatter.

* Add `days per file` item to model profile schema

This supports extractions from month-averaged datasets. The default value for
`days per file` is 1 so that it can be excluded from most model profiles
because they typically describe collections of files that contain 1 day of model
results per file.

* Add SalishSeaCast-201905-month-avg-salish model profile

* Add SalishSeaCast-202111-month-avg-salish model profile