Refactor variable encoding calculation
Reorganized the order of operations in calc_var_encoding() explicitly make time
coordinate chunk size the first element of the chunk sizes tuple. That is
necessary because the groupby() operation used to calculate climatology changes
the order of the items in the variable coordinates dict to put the time
coordinate last instead of first.

Also, added a new test to "tests/core/test_extract.py" to verify that the
changes remain compatible with functionality for extractions whether they
are climatology calculations or not.