Update unlimited dimension handling for climatology
Modified the calculation of the unlimited dimension in prep_netcdf_write()
to consider climatology datasets, where an unlimited time dimension doesn't make
sense. A conditional check for "climatology" in the configuration was
implemented. A test was also added in 'tests/core/test_extract .py' to ensure
the proper functioning of these changes.