Restore netcdf4 package for use as write engine
When h5netcdf is used as the engine for writes the resulting netCDF4 files are
not read correctly by ncdump or xarray.open_dataset(..., engine="netcdf4").
So, we'll use h5netcdf to read and netcdf4 to write.

Returning netcdf4 as a dependency also restores the ncdump tool to the
environment.