Update threads per worker to 1 in configs and docs (#143)
This is done to increase the reliability of Reshapr extractions. With threads>1
we see random occurrences of errors like:
  `RuntimeError: NetCDF: Not a valid ID`
The root cause appears to be that the `netcdf-c` is not thread-safe and a change
introduced in `netcdf4-python=1.6.1` removed a work-around for the lack of
thread-safety.

See discussion in https://github.com/xCDAT/xcdat/issues/561 and in other
discussions linked there.