Update to support Python 3.14 (#168)
* Add Python 3.14 to GHA pytest-with-coverage workflow

Expanded the `python-version` matrix in the workflow configuration to
include Python 3.14. Initially, this is to use the workflow to test
whether all the packages that Reshapr depends on have been updated to
support Python 3.14 on the conda-forge channel.

* Update environment descriptions to use Python 3.14

Upgraded Python from 3.13 to 3.14 across all YAML environment files. Also updated
the list of packages and versions used in the recent dev environment.

* Update sphinx-linkcheck workflow to Python 3.14

* Update docs to reflect Python 3.14 development env

The documentation is updated to specify Python 3.14 as the testing and development
environment, reflecting the project's support for the latest Python version. Minor
adjustments were also made to rendered outputs for consistency with the new
version.